+++ packages/components-native/src/InputFieldWrapper/components/ClearAction/index.ts
@@ -1,0 +1,2 @@
+export { ClearAction } from "./ClearAction";
+export { messages } from "./messages";